Refactored (moved) some methods around to different scripts.
