+++ tests/test_n_utils.py
@@ -5,10 +5,12 @@ import datetime
 import tests.initialize_env as init_env
 
 from pcse.input.nasapower import NASAPowerWeatherDataProvider
-from pcse_gym.utils.nitrogen_utils import (calculate_year_n_deposition,
-                                           convert_year_to_n_concentration,
-                                           calculate_day_n_deposition)
+from pcse_gym.utils.nitrogen_helpers import (calculate_year_n_deposition,
+                                             convert_year_to_n_concentration,
+                                             calculate_day_n_deposition)
 from pcse_gym.utils.weather_utils.weather_functions import generate_date_list
+from tests import initialize_env as init_env
+from utils.nitrogen_helpers import get_deposition_amount, get_disaggregated_deposition
 
 
 class TestNitrogenUtils(unittest.TestCase):
@@ -61,3 +63,47 @@ class TestNitrogenUtils(unittest.TestCase):
 
         self.assertAlmostEqual(day_nh4_depo, 0.0014, 1)
         self.assertAlmostEqual(day_no3_depo, 0.0025, 1)
+
+
+class NitrogenUseEfficiency(unittest.TestCase):
+    def setUp(self):
+        self.nue1 = init_env.initialize_env_nue_reward()
+        self.def1 = init_env.initialize_env_reward_dep()
+
+    def test_disaggregate(self):
+        start = datetime.date(2002, 1, 1)
+        end = datetime.date(2002, 2, 1)
+
+        nh4_depo, no3_depo = get_deposition_amount(2002)
+
+        daily_nh4 = nh4_depo / 365
+        daily_no3 = no3_depo / 365
+
+        expected_nh4 = daily_nh4 * 31
+        expected_no3 = daily_no3 * 31
+
+        nh4_dis, no3_dis = get_disaggregated_deposition(year=2002, start_date=start, end_date=end)
+
+        self.assertEqual(expected_nh4, nh4_dis)
+        self.assertEqual(expected_no3, no3_dis)
+
+    def test_nue_calc_in_other_rf(self, reward_func='DEP'):
+        self.env_rew = init_env.initialize_env(reward=reward_func, pcse_env=2)
+        self.env_rew.overwrite_year(2002)
+        self.env_rew.reset()
+        terminated = False
+
+        while not terminated:
+            _, _, terminated, _, infos = self.env_rew.step(np.array([1]))
+
+        self.assertAlmostEqual(0.58, max(list(infos['NUE'].values())), 0)
+
+    def test_nue_all_rf(self):
+        self.test_nue_calc_in_other_rf('DEF')
+        self.test_nue_calc_in_other_rf('HAR')
+        self.test_nue_calc_in_other_rf('NUE')
+        self.test_nue_calc_in_other_rf('FIN')
+        self.test_nue_calc_in_other_rf('NUP')
+        self.test_nue_calc_in_other_rf('DNU')
+        self.test_nue_calc_in_other_rf('END')
+        self.test_nue_calc_in_other_rf('GRO')
